+++ src/page-weekly-snapshot/WeeklySnapshotPage.tsx
@@ -5,8 +5,8 @@ import Loading from "../design-system/Loading";
 import { useLocaleDataState } from "../locale-data-context";
 import FacilitySummaries from "./FacilitySummaries";
 import ImpactProjectionChart from "./ImpactProjectionChart";
+import LocaleStatsTable from "./LocaleStatsTable";
 import LocaleSummary from "./LocaleSummary";
-import LocaleSummaryTable from "./LocaleSummaryTable";
 import SnapshotPage from "./SnapshotPage";
 import SystemWideProjectionChart from "./SystemWideProjectionChart";
 
@@ -30,7 +30,7 @@ const WeeklySnapshotPage: React.FC = () => {
               </SnapshotPage>
               <SnapshotPage header="System Snapshot" subheader>
                 <SystemWideProjectionChart />
-                <LocaleSummaryTable />
+                <LocaleStatsTable />
               </SnapshotPage>
               <FacilitySummaries />
             </WeeklySnapshotContainer>
